+++ crates/cubek-tile/src/tile/view/mod.rs
@@ -1,8 +1,10 @@
+pub mod accumulate;
 pub mod flat;
 pub mod masked;
 pub mod matrix;
 pub mod quant;
 
+pub use accumulate::*;
 pub use flat::*;
 pub use masked::*;
 pub use matrix::*;
